Fix for iOS Paste-from-Excel bug
Pasting matrices from Excel to Free42 on iOS didn't work right because
Excel on iOS uses CR as the line ender, while Free42 assumed lines
always end in LF or CRLF (and so it simply ignored CR). Changed the
logic to translate CRLF->LF and CR->LF, i.e. translate CR to LF, and if
a CR is immediately followed by an LF, drop that LF.